review feedback: move get_sql.
