+++ buckets-mdapi/src/sql.rs
@@ -180,16 +180,3 @@ fn post_timer_metrics<T>(
         .with_label_values(&[&method.as_str(), success])
         .observe(t);
 }
-
-pub fn get_sql(vnode: u64) -> String {
-    [
-        "SELECT id, owner, bucket_id, name, created, modified, content_length, \
-         content_md5, content_type, headers, sharks, properties \
-         FROM manta_bucket_",
-        &vnode.to_string(),
-        &".manta_bucket_object WHERE owner = $1 \
-          AND bucket_id = $2 \
-          AND name = $3",
-    ]
-    .concat()
-}
